pyproject.toml
[tool.poetry]
name = "mihomo_config_generator"
version = "0.1.0"
description = ""
authors = ["Bruce Zhang <zttt183525594@gmail.com>"]
readme = "README.md"
packages = [{ include = "generator" }]

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
loguru = "^0.7.3"
requests-unixsocket2 = "^0.4.2"
ruamel-yaml = "^0.18.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    "."
]

[tool.poetry.scripts]
mihomo_config_generator = "generator.script:main"
mihomo_speedtest = "generator.script:speedtest"
